#1 알고리즘 구조 설계

+++ metro/src/main/java/com/example/metro/map/entity/SubwayStation.java
@@ -13,7 +13,7 @@ public class SubwayStation {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long subwayStationId;
+    private int subwayStationId;
     private String StationName;
     private String lineName;
     private boolean isTransfer;
