+++ metro/src/main/java/com/example/metro/map/repository/SubwayStationRepository.java
@@ -3,5 +3,5 @@ package com.example.metro.map.repository;
 import com.example.metro.map.entity.SubwayStation;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SubwayStationRepository extends JpaRepository<Long, SubwayStation> {
+public interface SubwayStationRepository extends JpaRepository<Integer, SubwayStation> {
 }
