+++ metro/src/main/java/com/example/metro/map/repository/SubwayTrackRepository.java
@@ -3,5 +3,5 @@ package com.example.metro.map.repository;
 import com.example.metro.map.entity.SubwayTrack;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SubwayTrackRepository extends JpaRepository<Long, SubwayTrack> {
+public interface SubwayTrackRepository extends JpaRepository<Integer, SubwayTrack> {
 }
